Split versionInfo files to reduce need for rebuilds. HDF5 code should now automatically compress datasets if possible.
