+++ private/versioningForwards.hpp
@@ -21,5 +21,16 @@ namespace scatdb {
 		DLEXPORT_SDBR hash::HASH_p getHashOfVersionInfo(const versionInfo&);
 		DLEXPORT_SDBR versionInfo_p getLibVersionInfo();
 
+		/**
+		* \brief Provides information about the build environment during compilation.
+		*
+		* This function is designed to provide information on a compiler's
+		* build environment. It is a header function because it is designed
+		* to reflect the compiler variables of an external project's code.
+		*
+		* @param out The output stream that receives the information.
+		*/
+		DLEXPORT_SDBR void debug_preamble(const versionInfo &v, std::ostream &out = std::cerr);
+
 	}
 }
